finished

+++ Controllers/TripController.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using WebApplication4.Context;
 using WebApplication4.Dtos;
+using WebApplication4.Models;
 
 namespace WebApplication4.Controllers;
 
@@ -19,7 +20,7 @@ public class TripController : ControllerBase
     [HttpGet("trips")]
     public async Task<IActionResult> GetTripsAsync()
     {
-        return Ok(await _context.Trips.OrderBy(t=>t.DateFrom).ToListAsync());
+        return Ok(await _context.Trips.OrderByDescending(t=>t.DateFrom).ToListAsync());
     }
 
     [HttpDelete("clients/{id:int}")]
@@ -28,15 +29,54 @@ public class TripController : ControllerBase
         var client = await _context.Clients.FindAsync(id);
         if (client is null)
         {
-            return NotFound("No client with such given id exists");
+            return NotFound("No client with given id exists");
         }
 
+        var trips = await _context.Clients.Where(c => c.IdClient == id).SelectMany(c => c.ClientTrips).CountAsync();
+        if (trips > 0)
+        {
+            return Conflict("Can't delete client with registered trips");
+        }
         return Ok(await _context.Clients.Where(c => c.IdClient == id).ExecuteDeleteAsync() > 0);
     }
 
     [HttpPost(("trips/{id:int}/clients"))]
-    public async Task<IActionResult> assignClientToTripAsync([FromBody] ClientTripDto dto)
+    public async Task<IActionResult> AssignClientToTripAsync([FromBody] ClientTripDto dto, int id)
     {
+        var client = await _context.Clients.FindAsync(dto.Pesel);
+        if (client is null)
+        {
+            client = new Client()
+            {
+                FirstName = dto.FirstName,
+                LastName = dto.LastName,
+                Email = dto.Email,
+                IdClient = dto.IdClient,
+                ClientTrips = new List<ClientTrip>(),
+                Pesel = dto.Pesel,
+                Telephone = dto.Telephone
+            };
+            await _context.Clients.AddAsync(client);
+        }
+        var inTrip = await _context.ClientTrips.Where(c=>dto.IdClient == c.IdClient).Where(c=>c.IdTrip == id).ToListAsync();
+        if (inTrip.Count > 0)
+        {
+            return Conflict("Client already registered for given trip");
+        }
         
+        var trip = await _context.Trips.FindAsync(id);
+        if (trip is null)
+        {
+            return NotFound("No trip with given id exists");
+        }
+        DateTime now = DateTime.Now;
+        await _context.ClientTrips.AddAsync(new ClientTrip()
+        {
+            IdClient = client.IdClient,
+            IdTrip = trip.IdTrip,
+            PaymentDate = DateTime.Parse(dto.PaymentTime),
+            RegisteredAt = now
+        });
+        return Ok("inserted");
     }
 }
